+++ src/cljs/wil/core.cljs
@@ -26,22 +26,30 @@
 ;; r/atom
 
 (defonce session (r/atom {:page :home}))
-(defonce notes   (r/atom nil))
+
 (defonce params  (r/atom nil))
 (defonce others  (r/atom nil)) ;; 必要か？
 (defonce note    (r/atom ""))
 (defonce md      (r/atom "preview"))
+(defonce notes   (r/atom nil))
+(defonce date-count  (r/atom "boke")) ;; r/atom の必要ない。
 
-;; async
-;; (defonce ans (r/atom nil))
-;; (defonce goods-bads (r/atom ""))
+(reset! date-count "aho")
 
+;; Warning: Reactive deref not supported in lazy seq,
+;; it should be wrapped in doall
 (defn reset-notes!
   "get the notes from `/api/notes/:login`, set it in `notes` r/atom."
   []
   (GET (str "/api/notes/" js/login)
     {:handler #(reset! notes %)
      :error-handler (fn [^js/Event e] (js/alert (.getMessage e)))}))
+
+(defn reset-date-count!
+  []
+  (GET "/api/date-count"
+    {:handler #(reset! date-count %)
+     :error-hadler (fn [^js/Event e] (js/alert (.getMessage e)))}))
 
 ;;-------------------------
 ;; navbar
@@ -163,9 +171,8 @@
     ;; (js/alert (md->html (:note note)))
     [:section.section>div.container>div.content
      [:h2 (:login note) ", " (:date note)]
-     ; [:div {:dangerouslySetInnerHTML
-     ;       {:__html (md->html (:note note))}}]
-     [:div {:dangerousSetInnerHTML {:__html (md->html (:note note))}}]
+     [:div {:dangerouslySetInnerHTML
+            {:__html (md->html (:note note))}}]
      [:hr]
      [:div
       [:button.button.is-small
@@ -252,21 +259,28 @@
       {:handler #(js/alert (format-goods-bads %))
        :error-handler #(js/alert (str "error: get " uri))})))
 
+(defn cnt [date]
+  (-> (filter #(= date (:date %)) @date-count)
+      first
+      :count))
+
 (defn notes-component []
   (fn []
     [:div
      [:ol
-      (for [[i note] (reverse (map-indexed vector @notes))]
-        [:p
-         {:key i}
-         [:button.button.is-warning.is-small
-          {:on-click (fn [_]
-                       (fetch-others! (:date note))
-                       (swap! session assoc :page :others))}
-          (str (:date note))]
-         note
-         [:a {:href (str "/#/my/" (:id note))}
-          (-> (:note note) str/split-lines first)]])]]))
+      (doall (for [[i note] (reverse (map-indexed vector @notes))]
+               [:p
+                {:key i}
+                [:button.button.is-warning.is-small
+                 {:on-click (fn [_]
+                              (fetch-others! (:date note))
+                              (swap! session assoc :page :others))}
+                 (str (:date note))]
+                " ("
+                (cnt (:date note))
+                ") "
+                [:a {:href (str "/#/my/" (:id note))}
+                 (-> (:note note) str/split-lines first)]]))]]))
 
 (defn done-todays?
   []
@@ -375,7 +389,7 @@
 ;; Initialize app
 (defn ^:dev/after-load mount-components []
   (rdom/render [#'navbar] (.getElementById js/document "navbar"))
-  (rdom/render [#'page] (.getElementById js/document "app")))
+  (rdom/render [#'page]   (.getElementById js/document "app")))
 
 ;18
 ;(defn ^:dev/after-load mount-components []
@@ -388,4 +402,5 @@
   (ajax/load-interceptors!)
   (hook-browser-navigation!)
   (reset-notes!)
+  (reset-date-count!)
   (mount-components))